+++ CromIAM/src/main/scala/cromiam/webservice/WomtoolRouteSupport.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.server.Directives._
 import cromiam.cromwell.CromwellClient
 import cromiam.sam.SamClient
 
-trait WomtoolRouteSupport extends RequestSupport with EnabledUserSupport {
+trait WomtoolRouteSupport extends RequestSupport {
   // When this trait is mixed into `CromIamApiService` the value of `cromwellClient` is the reader (non-abort) address
   val cromwellClient: CromwellClient
   val samClient: SamClient
